upgrade bevy engine to 0.6

+++ src/systems/animate_sprite.rs
@@ -10,7 +10,7 @@ pub fn animate_sprite(
 
         if timer.finished() {
             let texture_atlas = texture_atlases.get(texture_atlas_handle).unwrap();
-            sprite.index = ((sprite.index as usize + 1) % texture_atlas.textures.len()) as u32;
+            sprite.index = ((sprite.index as usize + 1) % texture_atlas.textures.len()) as usize;
         }
     }
 }
